refactor: move currency to sub-dir
