+++ packages/sealed_currencies/lib/src/model/currency/currency.dart
@@ -1,7 +1,7 @@
 import "package:sealed_languages/sealed_languages.dart";
 
-import "../data/fiat_currencies.data.dart";
-import "../helpers/fiat_currency/fiat_currency_json.dart";
+import "../../data/fiat_currencies.data.dart";
+import "../../helpers/fiat_currency/fiat_currency_json.dart";
 
 part "submodels/crypto_currency.dart";
 part "submodels/fiat_currency.dart";
