+++ packages/sealed_currencies/lib/src/model/currency/submodels/crypto_currency.dart
