+++ packages/sealed_currencies/lib/src/model/currency/submodels/fiat_currency.dart
