+++ packages/sealed_currencies/test/src/model/currency/submodels/crypto_currency_test.dart
@@ -1,4 +1,4 @@
-import "package:sealed_currencies/src/model/currency.dart";
+import "package:sealed_currencies/src/model/currency/currency.dart";
 import "package:test/test.dart";
 
 void main() => group("$CryptoCurrency", () {
