+++ tools/bin/data.dart
@@ -1,4 +1,4 @@
-import "package:cli/generators/data_list_generator.dart";
+import "package:cli/generators/bool_getters_generator.dart";
 import "package:cli/utils/args_parser.dart";
 
 /// Usage: `dart run :data sealed_currencies`.
@@ -6,5 +6,5 @@ Future<void> main(List<String> args) {
   final package = ArgsParser(args).maybePackageName();
   if (package == null) throw ArgumentError("Package name should be provided.");
 
-  return const DataListGenerator().generate(package);
+  return const BoolGettersGenerator().generate(package);
 }
